fix(m15): Nodemailer nicht in Client-Bundle ziehen — pure deadlines-Modul (#168)
Vercel-Build schlug fehl: die Client-Split-View importiert (über
assignment-status.ts) nextDeadline aus acceptance/engine.ts. engine.ts
zieht transitiv den Mail-Transport (nodemailer) mit — Node-only —, sodass
webpack im Client-Bundle net/dns/fs nicht auflösen konnte.

Fix: reine SLA-Deadline-Mathematik (isShortNotice/getSLAWindows/
nextDeadline/DeadlineInput) nach src/lib/acceptance/deadlines.ts
ausgelagert (keine Admin-Client-/Mail-Abhängigkeit). engine.ts
re-exportiert diese Symbole (bestehende Imports + SSOT unverändert),
assignment-status.ts importiert nun aus dem reinen Modul.

Verifiziert: next build grün, tsc grün, next lint grün, vitest 1082 grün.
