+++ src/lib/dispatch/assignment-status.ts
@@ -27,7 +27,7 @@
  */
 
 import type { Enums } from "@/lib/types/database"
-import { nextDeadline, type DeadlineInput } from "@/lib/acceptance/engine"
+import { nextDeadline, type DeadlineInput } from "@/lib/acceptance/deadlines"
 
 type RideStatus = Enums<"ride_status">
